% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_msd_peak.R
\name{annotate_msd_peak}
\alias{annotate_msd_peak}
\title{Annotate peaks in MSD files}
\usage{
annotate_msd_peak(
  label,
  file,
  peaks,
  peakIdx = NA,
  theoMz = 0,
  charge = 1,
  peakBaseline = 0
)
}
\arguments{
\item{label}{Character, label of the peak}

\item{file}{Character, filepath to edit}

\item{peaks}{\code{MALDIquant::MassPeaks}, with corresponding peaks to file or a vector in form c(mz, intensity)}

\item{peakIdx}{Integer, index of peak in \code{peaks} or NA if mz and intensity are given directly}

\item{theoMz}{Numeric, theoretical mass of labeled peak (defaults to 0)}

\item{charge}{Integer, charge of peak (defaults to 1)}

\item{peakBaseline}{Numeric, intensity of baseline at peak position (not really sure what it does... defaults to 0)}
}
\value{
writes annoteted file directly to disk
}
\description{
Annotate peaks in MSD files
}
