% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{C99_3xFLAG}
\alias{C99_3xFLAG}
\title{Sequence of C99 with added 3xFLAG}
\format{
An object of class \code{character} of length 1.
}
\usage{
C99_3xFLAG
}
\description{
131 amino acids. Old sequence as used for reaction in Szaruga et. al., 2017, Cell. 
Contains two amino acid long linker region (GT)
}
\keyword{datasets}
