% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{C99_deltaT}
\alias{C99_deltaT}
\title{Sequence of C99 with added 3xFLAG and shorter linker region}
\format{An object of class \code{character} of length 1.}
\usage{
C99_deltaT
}
\description{
130 amino acids.
Removed T from linker region (see \code{C99_3xFLAG}) to shift masses away from contaminations at mass of AICD50-99 and Abeta1-38.
}
\keyword{datasets}
