% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{C99_full}
\alias{C99_full}
\title{Products of GSEC activity on C99 including modified formes}
\format{
An object of class \code{data.frame} with 1087 rows and 3 columns.
}
\usage{
C99_full
}
\description{
Includes all Abetas starting from 1:5, 11, 17 or 19 and ending in the range of 23:49 as well as AICDs and Species the latter each as [M+H]+ and [M+H2]+2 as well as modifications.
}
\keyword{datasets}
