% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{C99T_commonSpecies}
\alias{C99T_commonSpecies}
\title{Common products of GSEC activity on C99_deltaT}
\format{An object of class \code{data.frame} with 35 rows and 3 columns.}
\usage{
C99T_commonSpecies
}
\description{
Includes all Abetas starting from 1 or 11 and ending in the range of 37:49 as well as AICDs and Species the latter each as [M+H]+ and [M+H2]+2
}
\keyword{datasets}
