% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceEdit.R
\name{cutSequence}
\alias{cutSequence}
\title{Generate list of fragments from sequence}
\usage{
cutSequence(sequence, start = 1, end = nchar(sequence), prefix = "Frag")
}
\arguments{
\item{sequence}{character, amino acid sequence}

\item{start}{numeric vector, starting postions}

\item{end}{numeric vector, end positions}

\item{prefix}{character, prefix to construct name from. 
Names will be in the from "prefixStart_end"}
}
\value{
named list of sequence fragments
}
\description{
Generate list of fragments from sequence
}
\examples{
Ab42 <- "DAEFRHDSGYEVHHQKLVFFAEDVGSNKGAIIGLMVGGVVIA"
cutSequence(Ab42, 1, c(38, 40, 42), prefix = "Ab")

}
