% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggspectrum.R
\name{.convertSpec}
\alias{.convertSpec}
\title{Convert MALDIquant Spectra or peaks to data.frame}
\usage{
.convertSpec(spec, ID, plotIdx)
}
\arguments{
\item{spec}{MALDIquant::MassObject (spectra or peaks).}

\item{ID}{Character, name (ID) of spectra.}

\item{plotIdx}{Numeric, index in list of spectra.}
}
\value{

}
\description{
Convert MALDIquant Spectra or peaks to data.frame
}
