% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EphA4}
\alias{EphA4}
\title{Sequence of EphA4}
\format{An object of class \code{character} of length 1.}
\usage{
EphA4
}
\description{
86 amino acids.
Alternative Substrate of GSEC.
}
\keyword{datasets}
