% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ErbB4}
\alias{ErbB4}
\title{Sequence of ErbB4}
\format{An object of class \code{character} of length 1.}
\usage{
ErbB4
}
\description{
107 amino acids.
Alternative Substrate of GSEC.
}
\keyword{datasets}
