% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceEdit.R
\name{generate_assigndf}
\alias{generate_assigndf}
\title{Generate data.frame with sequence fragments as used by \code{assign_species}}
\usage{
generate_assigndf(sequencelist, noSubstrateName = "None",
  fragmentList = list(Ab = list(start = 1, end = 37:49, charge = 1), AICD
  = list(start = 49:50, end = nchar(sequencelist[[1]]), charge = 1:2),
  Substrate = list(start = 1, end = nchar(sequencelist[[1]]), charge =
  1:2)))
}
\arguments{
\item{sequencelist}{named list, sequences to process. Need to have the same character length for the default value of fragmentList to work}

\item{fragmentList}{named list, giving the name of the fragments and their start and end positions. See default values for an example.}
}
\value{
data.frame with columns Sequence, Species and mz
}
\description{
Generate data.frame with sequence fragments as used by \code{assign_species}
}
\examples{
generate_assigndf(C99_3xFLAG)

}
