% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_peakdf.R
\name{generate_peakdf}
\alias{generate_peakdf}
\title{generate peak static data.frame}
\usage{
generate_peakdf(spectra, pick_meth = "SuperSmoother",
  halfWindowSize = 10, SNR = 3, binpeaks = FALSE)
}
\arguments{
\item{spectra}{}

\item{pick_meth}{}

\item{halfWindowSize}{}

\item{SNR}{}

\item{binpeaks}{}
}
\value{
df containing peak statistics
}
\description{
generate peak static data.frame
}
