% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_resultdf.R
\name{generate_resultdf}
\alias{generate_resultdf}
\title{Generate result data.frame containing species found in supplied list}
\usage{
generate_resultdf(peak_df, speciesdf, tol, totalInt_varNames = c("Erb",
  "Neu", "Ab"), separateIdInto = c("Substrate", "Type", "Condition"),
  subsetcol = separateIdInto[1], sep = "_")
}
\arguments{
\item{peak_df}{Data.frame containing peak data as generate by \code{generate_peakdf}}

\item{speciesdf}{Data.frame, containing columns:
"Substrate"  Character, the substrate from which the species was generated (used as grouping variable )
"Species"    Character, the name of the species 
"mz"         Numeric, m/z of species.}

\item{tol}{Numeric, tolerance for assignment.}

\item{totalInt_varNames}{Character vector, names of variables to sum up for total intensity (e.g. part of product name: "Ab")}

\item{separateIdInto}{Character vector, names of new columns into which the ID varibales (e.g. sample name from instrument seperated by \code{sep})}

\item{subsetcol}{Character, column of \code{peak_df} used for subsetting (usually the substrate used in reaction.)}

\item{sep}{Character, seperation character for sample names in "ID" column}
}
\value{
data.frame only containing peak information which could be annotated by species given in \code{speciesdf}
}
\description{
Generate result data.frame containing species found in supplied list
}
\examples{
head(generate_resultdf(test_peak_df, test_Ablist, tol = 5))
}
