% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClosest.R
\name{getClostest}
\alias{getClostest}
\title{get clostest matches}
\usage{
getClostest(x, y, n = NA, tol = NA, unique = FALSE)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector to match x to.}

\item{n}{Numeric, maximum number of matches}

\item{tol}{Numeric, tolerance for matching.}

\item{unique}{Logical, only retrun unique matches (if n>1).}
}
\description{
get clostest matches
}
