% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceEdit.R
\name{getMassOfSequece}
\alias{getMassOfSequece}
\title{Get the mass of ions from sequence}
\usage{
getMassOfSequece(sequence, charge = 1, mode = "avg", IAA = FALSE,
  unlist = TRUE)
}
\arguments{
\item{sequence}{character or (named) character vector, amino acid sequence}

\item{charge}{numeric, number of charges of the ion}

\item{mode}{character, can be "mono" (monoisotopic mass) or "avg" (average mass)}

\item{IAA}{logical, see \code{OrgMassSpecR::ConvertPeptide}}

\item{unlist}{logical, unlist result to return (named) vector of masses.}
}
\value{
list or single (named) numeric value
}
\description{
Caution:
The returned mass differs at the second decimal position from the values returned by Bruker's SequenceEditor's Charge states function!
This is no big deal for linear mode and should also be ok for most reflector mode applications but can not be used to determine exact masses for FT-ICR.
}
\examples{

getMassOfSequece(C99_3xFLAG)
}
