% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNormFactors.R
\name{getNormFactors}
\alias{getNormFactors}
\title{Get normalization factors for a specific species}
\usage{
getNormFactors(peaksdf, speciesdf, targetSpecies, tol, tolppm = TRUE,
  allowNoMatch = TRUE)
}
\arguments{
\item{peaksdf}{Data.frame, see \code{AssaySupport::generate_peakdf}.}

\item{speciesdf}{Data.frame, see \code{AssaySupport::generate_assigndf}.}

\item{targetSpecies}{Character, name of species for which the normalization factors should be extracted.}

\item{tol}{Numeric, tolerance in Dalton or (if \code{tolppm = TRUE}) ppm}

\item{tolppm}{Logical, use tolerance in ppm.}

\item{allowNoMatch}{Logical, exclude spectra where \code{targetSpecies} was not found.}
}
\value{
Numeric vector of normalization factors.
}
\description{
Get normalization factors for a specific species
}
