% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{ggoverlay_spectra}
\alias{ggoverlay_spectra}
\title{Plot spectrum with ggplot}
\usage{
ggoverlay_spectra(spec_ID, labels_df = label_df, specs = spectra_align,
  xlim = c(4000, 5000), n_overlay = 3, showLabel = T,
  normalizing = "max", tol = 3)
}
\arguments{
\item{spec_ID}{}

\item{labels_df}{}

\item{specs}{}

\item{xlim}{}

\item{n_overlay}{}

\item{showLabel}{}

\item{normalizing}{}

\item{tol}{}
}
\value{
ggplot
}
\description{
Plot spectrum with ggplot
}
