% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{ggoverview}
\alias{ggoverview}
\title{Overview using ggplot}
\usage{
ggoverview(spec_ID, n_overlay, labels_df = label_df,
  specs = spectra_align, ab_max = get_max(labels_df, "Ab"),
  aicd_max = get_max(labels_df, "AICD"), label = "auto",
  total_xlim = c(3000, 16000), abeta_xlim = c(4000, 5000),
  aicd_xlim = c(9400, 10000), substr_xlim = c(14500, 15500))
}
\arguments{
\item{spec_ID}{}

\item{n_overlay}{}

\item{labels_df}{}

\item{specs}{}

\item{ab_max}{}

\item{aicd_max}{}
}
\value{
ggplot
}
\description{
Overview using ggplot
}
