% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_spectra.R
\name{load_spectra}
\alias{load_spectra}
\title{load spectra of the bruker flex format}
\usage{
load_spectra(Dir, filter = c("Cal", "cal", "CAL"), nameSpectra = TRUE)
}
\arguments{
\item{Dir}{Character, directory containing spectra}

\item{filter}{Character vector containing spectra names to be filtered}

\item{nameSpectra}{Logical, name spectra according to there parent folder (default)}
}
\value{
list of MALDIquant spectra
}
\description{
load spectra of the bruker flex format
}
