% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_multiple.R
\name{match_mutiple}
\alias{match_mutiple}
\title{Relaxed value matching with multiple results}
\usage{
match_mutiple(x, vec, tol = Inf, n = 1, tol_ppm = FALSE)
}
\arguments{
\item{x}{numeric, value to match}

\item{vec}{numeric vector, to match \code{x} against}

\item{tol}{numeric, tolerance for matching. 
Multiply with 1e-6 and set \code{tol_ppm} to use relative tolerance.}

\item{n}{numeric, maximal number of matches. Set to "all" for all matches within \code{tol}}

\item{tol_ppm}{logical, set to \code{TRUE} if \code{tol} should be used as ppm.}
}
\value{
Integer vector, indeces of \code{vec} matching \code{x}.
}
\description{
Relaxed value matching with multiple results
}
