% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{mutation_plot}
\alias{mutation_plot}
\title{Plot comparing different mutations using base R graphics}
\usage{
mutation_plot(idx, plot_n = 4, specs = spectra_align,
  labeldf = label_df, Ab_xlim = c(3500, 5500), ymax_label = TRUE,
  ymax_adjust = 1.7)
}
\arguments{
\item{idx}{}

\item{plot_n}{}

\item{specs}{}

\item{labeldf}{}

\item{Ab_xlim}{}

\item{ymax_label}{}

\item{ymax_adjust}{}
}
\value{
plot
}
\description{
Plot comparing different mutations using base R graphics
}
