% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCadherin}
\alias{NCadherin}
\title{Sequence of N-Cadherin}
\format{An object of class \code{character} of length 1.}
\usage{
NCadherin
}
\description{
109 amino acids.
Alternative Substrate of GSEC.
}
\keyword{datasets}
