% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Neurexin}
\alias{Neurexin}
\title{Sequence of Neurexin}
\format{
An object of class \code{character} of length 1.
}
\usage{
Neurexin
}
\description{
121 amino acids.
Alternative Substrate of GSEC.
}
\keyword{datasets}
