% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeByFactor.R
\name{normalizeByFactor}
\alias{normalizeByFactor}
\title{Normalize spectra by factor}
\usage{
normalizeByFactor(spec, factors)
}
\arguments{
\item{spec}{list, list of \code{MassSpectrum} objects.}

\item{factors}{numeric vector, scaling factors. Needs to have same length as \code{spec}.}
}
\value{
Scaled spectra
}
\description{
Normalize spectra by factor
}
