% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{overviewplot}
\alias{overviewplot}
\title{Overview using base R graphics}
\usage{
overviewplot(idx, specs = spectra_align, labeldf = label_df,
  total_xlim = c(3000, 16000), AICD_xlim = c(9500, 10000),
  Ab_xlim = c(3500, 5500), subst_xlim = c(13500, 15500),
  offset = 1.9, ymax_label = TRUE, ymax_adjust = 1.7,
  overview_adjust = 5)
}
\arguments{
\item{idx}{}

\item{specs}{}

\item{labeldf}{}

\item{total_xlim}{}

\item{AICD_xlim}{}

\item{Ab_xlim}{}

\item{subst_xlim}{}

\item{offset}{}

\item{ymax_label}{}

\item{ymax_adjust}{}

\item{overview_adjust}{}
}
\value{
plot
}
\description{
Overview using base R graphics
}
