% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PAP_cal}
\alias{PAP_cal}
\title{Calibration spectrum containing average masses of:
name				                avg		    mono
ACTH_clip(18-39)		        2466.68		
ACTH_clip(1-17)		        2094.42				
Somatostatin			          3149.574
Ab1-28				              3263.472
Ab1-38				              4131.548	4132.5557
Ab1-39 				            4230.680	4231.6870
Ab1-40 				            4329.811	4330.8182
Ab1-42 				            4514.047	4515.0541
Ab1-43 				            4615.151	4616.1582
Ab1-45 				            4827.440	4828.4473
Ab1-46 				            4926.571	4927.5785
Ab1-48 				            5140.833	5141.8405
Insulin_[M+H]+_avg 		    5734.520
Cytochrome_C_[M+2H]2+_avg 	6181.050
Myoglobin_[M+2H]2+_avg 		8476.660
Ubiquitin_I_[M+H]+_avg 		8565.760}
\format{
An object of class \code{MassPeaks} of length 16.
}
\usage{
PAP_cal
}
\description{
Calibration spectrum containing average masses of:
name				                avg		    mono
ACTH_clip(18-39)		        2466.68		
ACTH_clip(1-17)		        2094.42				
Somatostatin			          3149.574
Ab1-28				              3263.472
Ab1-38				              4131.548	4132.5557
Ab1-39 				            4230.680	4231.6870
Ab1-40 				            4329.811	4330.8182
Ab1-42 				            4514.047	4515.0541
Ab1-43 				            4615.151	4616.1582
Ab1-45 				            4827.440	4828.4473
Ab1-46 				            4926.571	4927.5785
Ab1-48 				            5140.833	5141.8405
Insulin_[M+H]+_avg 		    5734.520
Cytochrome_C_[M+2H]2+_avg 	6181.050
Myoglobin_[M+2H]2+_avg 		8476.660
Ubiquitin_I_[M+H]+_avg 		8565.760
}
\keyword{datasets}
