% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceEdit.R
\name{pointMutateSequence}
\alias{pointMutateSequence}
\title{Change single amino acid in sequence}
\usage{
pointMutateSequence(sequence, pos, substitute, names = NULL)
}
\arguments{
\item{sequence}{Character, sequence to edit.}

\item{pos}{Integer vector, position at which \code{sequence} should be edited.}

\item{substitute}{Character vector, amino acid (sequece) to change the amino acid at \code{pos} with.}

\item{names}{Character vector, names of generated sequences. 
Set to \code{NULL} to generate systemic names (e.g. D7N for Tottori)}
}
\value{
mutated amino acid sequences
}
\description{
Change single amino acid in sequence
}
\examples{
pointMutateSequence(sequence = C99_deltaT, 
                    pos = C99_mutations$pos, 
                    substitute = C99_mutations$substitute, 
                    names = C99_mutations$name)
                    
}
