% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profilePlot.R
\name{profilePlot}
\alias{profilePlot}
\title{Product profile plot}
\usage{
profilePlot(
  resultdf,
  group_var,
  select_species = "Ab",
  filter = "IS",
  show_legend = FALSE
)
}
\arguments{
\item{resultdf}{Data.frame as generated by \code{generate_resultdf}}

\item{group_var}{Character, colname in resultdf used for grouping (x-axis of the resulting plot).}

\item{select_species}{Character, the name of the target analyte. 
This is used in a grepl so it can also be a fragment of the full name (e.g. Ab for Abx_y).}

\item{filter}{Character, names to filter out.
This is used in a grepl so it can also be a fragment of the full name (e.g. IS for IS_Ab1_28).}

\item{show_legend}{Logical, show legend in addition to labels.}
}
\value{
ggplot object
}
\description{
Generates a product profile plot.
Uses a stacked bar plot on percentage values normalized to total product intensity 
to visulize shifts in the profile.
}
