% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readScilsCsv.R
\name{readScilsCsv}
\alias{readScilsCsv}
\title{Read csv of overview spectra from SCiLS}
\usage{
readScilsCsv(path, type = c("mean", "skyline", "variance"))
}
\arguments{
\item{path}{Character, path to folder with .csv files}

\item{type}{Character, type of spectrum to import.}
}
\value{
List of MALDIquant::MassSpectrum
}
\description{
Read csv of overview spectra from SCiLS
}
