% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftMassAxis.R
\name{shiftMassAxis}
\alias{shiftMassAxis}
\title{Shift mass axis/Linear point correction}
\usage{
shiftMassAxis(spec, mzdiff)
}
\arguments{
\item{spec}{MALDIquant::MassSpectrum or MALDIquant::MassPeaks}

\item{mzdiff}{numeric, correction value for mass axis.}
}
\value{
MALDIquant::MassSpectrum or MALDIquant::MassPeaks with shifted mass axis
}
\description{
Shift mass axis/Linear point correction
}
