% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggspectrum.R
\name{spec2df}
\alias{spec2df}
\title{Convert MALDIquant::MassSpectrum to tibble}
\usage{
spec2df(spec, peaks, nameFromMetaData = FALSE)
}
\arguments{
\item{spec}{(list of) MALDIquant::MassSpectrum}

\item{peaks}{(list of) MALDIquant::MassPeaks. Optinal.}

\item{nameFromMetaData}{Logical, get names from metaData or from named list.}
}
\value{
tibble with spectra and peak data.
}
\description{
Convert MALDIquant::MassSpectrum to tibble
}
