% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_peaks}
\alias{test_peaks}
\title{Peaks of (\code{test_spectra_proc}) as \code{MALDIquant::MassPeaks}}
\format{
An object of class \code{list} of length 2.
}
\usage{
test_peaks
}
\description{
Peaks of (\code{test_spectra_proc}) as \code{MALDIquant::MassPeaks}
}
\keyword{datasets}
