% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_result_df}
\alias{test_result_df}
\title{Example of \code{result_df} containing annotations}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5 rows and 13 columns.}
\usage{
test_result_df
}
\description{
Example of \code{result_df} containing annotations
}
\keyword{datasets}
