% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_spectra_proc}
\alias{test_spectra_proc}
\title{Preprocessed spectra (\code{MALDIquant::MassSpectrum}) of GSEC reactions as examples}
\format{
An object of class \code{list} of length 2.
}
\usage{
test_spectra_proc
}
\description{
Preprocessed spectra (\code{MALDIquant::MassSpectrum}) of GSEC reactions as examples
}
\keyword{datasets}
