% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector2peakdf.R
\name{vector2peakdf}
\alias{vector2peakdf}
\title{Convert vector of m/z values to a peakdf as used by many function}
\usage{
vector2peakdf(mz, ID = "C99_0_0")
}
\arguments{
\item{mz}{Numerical, mz values}

\item{ID}{Character, ID as created by FlexControl in format "SUBSTRATE_TYPE_CONDITION"}
}
\value{
peakdf
}
\description{
Convert vector of m/z values to a peakdf as used by many function
}
