% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_msd}
\alias{write_msd}
\title{write spectra as MSD to disk}
\usage{
write_msd(spectra = spectra_align, parentDir = parentDir,
  namePrefix = namePrefix)
}
\arguments{
\item{spectra}{}

\item{parentDir}{}

\item{namePrefix}{}
}
\description{
write spectra as MSD to disk
}
