% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_peak_list}
\alias{write_peak_list}
\title{write peak statistics as xlsx file to disk}
\usage{
write_peak_list(df, dir, namePrefix = "")
}
\arguments{
\item{df}{Data.frame, as generated by \code{generate_resultdf()}}

\item{dir}{Character, directory to write into}

\item{namePrefix}{Character, prefix to add to filename}
}
\description{
write peak statistics as xlsx file to disk
}
