% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_species.R
\name{assign_species}
\alias{assign_species}
\title{Assign species to mz values}
\usage{
assign_species(peak_df, mzlist, subsetcol = "Substrate", tol = 5,
  speciescol = "Species", mzcol = "mz", append = TRUE)
}
\arguments{
\item{peak_df}{}

\item{mzlist}{}

\item{tol}{}

\item{append}{}

\item{subset.col}{}

\item{species.col}{}

\item{mz.col}{}
}
\value{
df with assigned peaks if append is TRUE
}
\description{
Assign species to mz values
}
