% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_species.R
\name{assign_species}
\alias{assign_species}
\title{Assign species to mz values}
\usage{
assign_species(peak_df, mzlist, subsetcol = "Substrate", tol = 5,
  speciescol = "Species", mzcol = "mz", append = TRUE)
}
\arguments{
\item{peak_df}{n}

\item{mzlist}{n}

\item{tol}{n}

\item{append}{n}

\item{subset.col}{n}

\item{species.col}{n}

\item{mz.col}{n}
}
\value{
df with assigned peaks if append is TRUE
}
\description{
Assign species to mz values
}
