% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_species.R
\name{assign_species}
\alias{assign_species}
\title{Assign species to mz values}
\usage{
assign_species(
  peak_df,
  speciesdf,
  subsetcol = "Substrate",
  highMz = NA,
  tol = 5,
  highMzTol = NA,
  tolppm = FALSE,
  speciescol = "Species",
  mzcol = "mz",
  append = TRUE
)
}
\arguments{
\item{peak_df}{Data.frame, as generated by \code{generate_peakdf()}}

\item{speciesdf}{Data.frame, containing columns:
        "Substrate"  Character, the substrate from which the species was generated (used as grouping variable )
        "Species"    Character, the name of the species 
        "mz"         Numeric, m/z of species.
Column names can be changed but have to be specified in thid case by \code{subset.col, species.col, mz.col}}

\item{subsetcol}{Character, name of column in \code{speciesdf} to subset data by (eg. "Substrate")}

\item{highMz}{Numeric, if provided a second tolerance (\code{highMzTol}) for \code{mz >= highMz} will be used.}

\item{tol}{Numeric, tolerance for assignment either in Da or ppm.}

\item{highMzTol}{Numeric, tolerance for assignment for \code{mz >= highMz} either in Da or ppm.}

\item{tolppm}{Logical, use relative tolerance in ppm (\code{tol} is multiplied by 1e-6 internally).}

\item{speciescol}{Character, name of column in \code{speciesdf} containing names of species}

\item{mzcol}{Character, name of column in \code{speciesdf} containing m/z of species}

\item{append}{Logical, if \code{TRUE} the result is appended as additional columns to the origninal data. 
Otherwise only the resulting assignments are returned.}
}
\value{
Data.frame with assigned peaks if append is TRUE
}
\description{
Assign species to mz values
}
