% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_peakdf.R
\name{generate_peakdf}
\alias{generate_peakdf}
\title{generate peak statistic data.frame}
\usage{
generate_peakdf(
  data,
  pick_meth = "SuperSmoother",
  halfWindowSize = 10,
  SNR = 3,
  binpeaks = FALSE,
  binTol = 5e-05
)
}
\arguments{
\item{data}{\code{MALDIquant::MassSpectrum} or \code{MALDIquant::MassPeaks}}

\item{pick_meth}{Character, peak picking method. Only applies is no peaks are supplied.}

\item{halfWindowSize}{Integer, halfWindowSize. Only applies is no peaks are supplied.}

\item{SNR}{numeric, SNR value. Only applies is no peaks are supplied.}

\item{binpeaks}{Logical, perform binning.}
}
\value{
data.frame containing peak statistics
}
\description{
generate peak statistic data.frame
}
\examples{
head(generate_peakdf(test_spectra_proc))
}
