% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_peakdf.R
\name{generate_peakdf}
\alias{generate_peakdf}
\title{generate peak static data.frame}
\usage{
generate_peakdf(spectra, pick_meth = "SuperSmoother",
  halfWindowSize = 10, SNR = 3, binpeaks = FALSE)
}
\arguments{
\item{spectra}{spectra}

\item{pick_meth}{peak picking method}

\item{halfWindowSize}{halfWindowSize}

\item{SNR}{SNR value}

\item{binpeaks}{logical, bin peaks?}
}
\value{
df containing peak statistics
}
\description{
generate peak static data.frame
}
