% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_peakdf.R
\name{generate_peakdf}
\alias{generate_peakdf}
\title{generate peak statistic data.frame}
\usage{
generate_peakdf(spectra, pick_meth = "SuperSmoother",
  halfWindowSize = 10, SNR = 3, binpeaks = FALSE, binTol = 5e-05)
}
\arguments{
\item{spectra}{\code{MALDIquant::MassSpectrum}}

\item{pick_meth}{Character, peak picking method}

\item{halfWindowSize}{Integer, halfWindowSize}

\item{SNR}{numeric, SNR value}

\item{binpeaks}{Logical, perform binning}
}
\value{
data.frame containing peak statistics
}
\description{
generate peak statistic data.frame
}
\examples{
head(generate_peakdf(test_spectra_proc))
}
