% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{get_max}
\alias{get_max}
\title{get max value of df subsetted by x}
\usage{
get_max(labels_df, x)
}
\arguments{
\item{labels_df}{}

\item{x}{}
}
\value{
max value
}
\description{
get max value of df subsetted by x
}
