% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMzShift.R
\name{getMzShift}
\alias{getMzShift}
\title{Get normalization factors for a specific species}
\usage{
getMzShift(peaksdf, tol, targetSpecies, speciesdf = NA, tolppm = TRUE,
  allowNoMatch = TRUE)
}
\arguments{
\item{peaksdf}{Data.frame, see \code{AssaySupport::generate_peakdf}.}

\item{tol}{Numeric, tolerance in Dalton or (if \code{tolppm = TRUE}) ppm}

\item{targetSpecies}{Character or Numeric, name of species for which the normalization factors should be extracted.
Or, if numeric, mz of species used for normalization.}

\item{speciesdf}{Data.frame, see \code{AssaySupport::generate_assigndf}.}

\item{tolppm}{Logical, use tolerance in ppm.}

\item{allowNoMatch}{Logical, exclude spectra where \code{targetSpecies} was not found.}
}
\value{
Numeric vector of normalization factors.
}
\description{
Get normalization factors for a specific species
}
