% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggspectrum.R
\name{ggSpectrum}
\alias{ggSpectrum}
\title{ggplot wrapper to plot spectra}
\usage{
ggSpectrum(df, minSNR = 1, scale = 100, round = 1)
}
\arguments{
\item{df}{tibble, as generated by spec2df()}

\item{minSNR}{numeric, filter peaks with snr lower then minSNR.}

\item{scale}{numeric or character, either numeric to multiply intensity or "min" or "max" to scale spectra to specified measure.}

\item{round}{integer, number of digits to display in plot.}
}
\value{
ggplot object.
}
\description{
ggplot wrapper to plot spectra
}
