% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_multiple.R
\name{match_mutiple}
\alias{match_mutiple}
\title{Relaxed value matching with multiple results}
\usage{
match_mutiple(x, vec, tol = Inf, n = 1)
}
\arguments{
\item{x}{numeric, value to match}

\item{vec}{numeric vector, to match \code{x} against}

\item{tol}{numeric, tolerance for matching}

\item{n}{numeric, maximal number of matches. Set to "all" for all matches within \code{tol}}
}
\value{
Integer vector, indeces of \code{vec} matching \code{x}.
}
\description{
Relaxed value matching with multiple results
}
