% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{mutation_plot}
\alias{mutation_plot}
\title{Plot comparing different mutations using base R graphics}
\usage{
mutation_plot(idx, plot_n = 4, specs = spectra_align,
  labeldf = label_df, Ab_xlim = c(3500, 5500), ymax_label = TRUE,
  ymax_adjust = 1.7)
}
\arguments{
\item{idx}{n}

\item{plot_n}{n}

\item{specs}{n}

\item{labeldf}{n}

\item{Ab_xlim}{n}

\item{ymax_label}{n}

\item{ymax_adjust}{n}
}
\value{
plot
}
\description{
Plot comparing different mutations using base R graphics
}
