% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{overviewplot}
\alias{overviewplot}
\title{Overview using base R graphics}
\usage{
overviewplot(idx, specs = spectra_align, labeldf = label_df,
  total_xlim = c(3000, 16000), AICD_xlim = c(9500, 10000),
  Ab_xlim = c(3500, 5500), subst_xlim = c(13500, 15500),
  offset = 1.9, ymax_label = TRUE, ymax_adjust = 1.7,
  overview_adjust = 5)
}
\arguments{
\item{idx}{index}

\item{specs}{spectra}

\item{labeldf}{df with labels}

\item{total_xlim}{limits}

\item{AICD_xlim}{limits}

\item{Ab_xlim}{limits}

\item{subst_xlim}{limits}

\item{offset}{offset}

\item{ymax_label}{n}

\item{ymax_adjust}{n}

\item{overview_adjust}{n}
}
\value{
plot
}
\description{
Overview using base R graphics
}
