% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_spectra}
\alias{preprocess_spectra}
\title{Preprocess spectra}
\usage{
preprocess_spectra(spec = spectra, smooth_meth = "SavitzkyGolay",
  smooth_halfWindowSize = 20, norm_meth = "TIC",
  filter_spectra = "Cal|D1|D2|CAL", baseline_meth = "TopHat",
  avg_method = "mean", align = c("none", "before", "after", "both"),
  alignMethod = "linear")
}
\arguments{
\item{spec}{List of \code{MALDIquant::MassSpectrum}}

\item{smooth_meth}{Character, smoothing method see \code{MALDIquant::smoothIntensity()}}

\item{smooth_halfWindowSize}{Numeric, halfWindowSize for smoothing method see \code{MALDIquant::smoothIntensity()}}

\item{norm_meth}{Character, normalization method see \code{MALDIquant::calibrateIntensity()}}

\item{filter_spectra}{Character vector, regex of patterns to exclude spectra (like calibration spectra in same folder)}

\item{baseline_meth}{Character, baseline removal method see \code{MALDIquant::removeBaseline}}

\item{avg_method}{Character, aggregation method used to generate average spectra. See \code{MALDIquant::averageMassSpectra}.}

\item{align}{Chacter vector, perform alignment before, after or both, before and after aggregation.}

\item{alignMethod}{Character, Alignment method (see \code{MALDIquant::alignSpectra}.)}
}
\value{
List of preprocessed \code{MALDIquant::MassSpectrum}
}
\description{
Preprocess spectra
}
\examples{
test_spectra_proc <- AssaySupport::preprocess_spectra(spec = test_spectra,
                                                      smooth_meth = "SavitzkyGolay",
                                                      smooth_halfWindowSize  = 5,
                                                      norm_meth = "TIC",
                                                      filter_spectra = NA,
                                                      baseline_meth = "TopHat",
                                                      align = "none")

}
