% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_spectra}
\alias{preprocess_spectra}
\title{Preprocess spectra}
\usage{
preprocess_spectra(spec = spectra, smooth_meth = "SavitzkyGolay",
  smooth_halfWindowSize = 20, norm_meth = "TIC",
  filter_spectra = "Cal|D1|D2|CAL", baseline_meth = "TopHat",
  align = FALSE)
}
\arguments{
\item{smooth_meth}{}

\item{smooth_halfWindowSize}{}

\item{norm_meth}{}

\item{filter_spectra}{}

\item{baseline_meth}{}

\item{spectra}{}
}
\value{
preprocessed spectra
}
\description{
Preprocess spectra
}
