% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_spectra}
\alias{preprocess_spectra}
\title{Preprocess spectra}
\usage{
preprocess_spectra(spec = spectra, smooth_meth = "SavitzkyGolay",
  smooth_halfWindowSize = 20, norm_meth = "TIC",
  filter_spectra = "Cal|D1|D2|CAL", baseline_meth = "TopHat",
  align = FALSE)
}
\arguments{
\item{smooth_meth}{Character, smoothing method see \code{MALDIquant::smoothIntensity()}}

\item{smooth_halfWindowSize}{Numeric, halfWindowSize for smoothing method see \code{MALDIquant::smoothIntensity()}}

\item{norm_meth}{Character, normalization method see \code{MALDIquant::calibrateIntensity()}}

\item{filter_spectra}{Character vector, regex of patterns to exclude spectra (like calibration spectra in same folder)}

\item{baseline_meth}{Character, baseline removal method see \code{MALDIquant::removeBaseline}}

\item{spectra}{List of \code{MALDIquant::MassSpectrum}}
}
\value{
List of preprocessed \code{MALDIquant::MassSpectrum}
}
\description{
Preprocess spectra
}
