% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggspectrum.R
\name{spec2df}
\alias{spec2df}
\title{Convert MALDIquant::MassSpectrum to tibble}
\usage{
spec2df(spec, peaks, result_df, nameFromMetaData = FALSE, unlist = TRUE)
}
\arguments{
\item{spec}{(list of) MALDIquant::MassSpectrum}

\item{peaks}{(list of) MALDIquant::MassPeaks. Optinal.}

\item{result_df}{Data.frame, annotations as generated by \code{AssaySupport::generate_resultdf}.}

\item{nameFromMetaData}{Logical, get names from metaData or from named list.}

\item{unlist}{Logical, unlist result.}
}
\value{
tibble with spectra and peak data.
}
\description{
Convert MALDIquant::MassSpectrum to tibble
}
