% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_peak_list}
\alias{write_peak_list}
\title{write peak statistics as xlsx file to disk}
\usage{
write_peak_list(labels_df = label_df, parentDir = parentDir,
  namePrefix = namePrefix)
}
\arguments{
\item{labels_df}{}

\item{parentDir}{}

\item{namePrefix}{}
}
\description{
write peak statistics as xlsx file to disk
}
